test coverage for the statistics and preprocess completed
